Port modules account_banking_fr_lcr and l10n_fr_siret to v10
Prepare the migration of l10n_fr_naf_ape to v10
